boyscout: move const declaration closer to usage; clean up unneeded code
